refactor(ast_visit): split UTF-8 to UTF-16 converter into multiple files (#13236)
Pure refactor. Split up the converter for UTF-8 -> UTF-16 offsets into multiple files.

Adding support for `loc` (#10307) will require adding more code to the converter. Splitting it up into logical chunks makes the code easier to navigate and understand.
